requirements.txt
requests
loguru
dashscope
haystack-ai
colbert-ai
rich
gradio
gradio_client
pandasai
sentence-transformers